refactor(fleetnode): fold detectLocalSubnets into nmap.go
After the combined-mode parity change, localsubnet.go held only errNoLocalSubnet
and a small detectLocalSubnets wrapper used solely by buildNmapOptions. Move both
next to their caller in nmap.go and drop the now-empty file.
